+++ apps/frontend/src/_data/user.ts
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 import { z } from 'zod';
 import { fetchUserInfo } from '@/_apis/user';
 import { UserInfoResponseSchema } from '@repo/schema/user';
-import { APIError } from '@/_apis/fetcher';
+import { APIError } from '@/_libs/error/errors';
 
 type UserInfoResponse = z.infer<typeof UserInfoResponseSchema>;
